+++ CHANGELOG.md
@@ -201,6 +201,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fix docblocks
     - [PR-815](https://github.com/OXID-eSales/oxideshop_ce/pull/815)
     - [PR-816](https://github.com/OXID-eSales/oxideshop_ce/pull/816)
+- Fix Shop deletes savedbasket from all subshops when discarding reservation basket [#0007136](https://bugs.oxid-esales.com/view.php?id=7136)
 
 ### Changed
 - jQuery UI updated to v1.12.1